Fixed loading of images in cubext.
The problem for loading images is not completly in my code. It sure
didn't help me much but the real problem is located inside the

binding for the chicken egg : soil.

It is creating a blob of size (* width height (sizeof char)). It is
completely ignoring the byte per pixel that is necessary to get every
bytes for each pixel. A standard RGB image has 3 byte per pixel, while
some image format may add a fourth channel like RGBA with 4 channel.

I'll have to push a patch to this egg to actually make sure that it get
fixed.